refactor(colors): rename color submodule to colors
Move src/dartwork_mpl/color/ to src/dartwork_mpl/colors/ and update
every internal import (src, tests, scripts, docs/api, docs/color_system).
No backward-compat shim — internal-only project, near-zero external
users. Also includes the implementation plan document.

Refs #164.
