+++ src/dartwork_mpl/colors/_color.py
