+++ src/dartwork_mpl/colors/_typing.py
