+++ src/dartwork_mpl/colors/_views.py
